+++ test/migrations-parse/20241012170200-create-quux-table.up.sql
@@ -1,0 +1,6 @@
+-- test comment parsed correctly
+CREATE TABLE -- first comment
+quux -- second comment;
+(id bigint,
+ name varchar(255)); -- last comment
+-- end