+++ test/migratus/mock.clj
@@ -18,6 +18,8 @@
   proto/Migration
   (id [this]
     id)
+  (migration-type [this]
+    :sql)
   (name [this]
     name)
   (up [this config]
@@ -32,6 +34,8 @@
   (init [this])
   (completed-ids [this]
     @completed-ids)
+  (completed [this]
+    (map (fn [id] {:id id :applied true}) @completed-ids))
   (migrate-up [this migration]
     (proto/up migration config)
     (swap! completed-ids conj (proto/id migration))